refactor: data loading, debounce and analytics leave utils.js
loadData with its node and browser branches and runningInNode move to
loader.js, debounce to debounce.js, and noteEvent to web/analytics.js, so
importing a core module no longer reaches for gtag. Tests that spied on the
loader through the utils namespace spy on loader.js.

Part of #1072.
