+++ src/mmc.js
@@ -1,5 +1,5 @@
-import { loadData } from "./utils.js";
 import { createZipBlob, unzip } from "./archive.js";
+import { loadData } from "./loader.js";
 
 /*
 highly adapted from:
