+++ src/web/google-drive-picker.js
@@ -1,10 +1,10 @@
-import * as utils from "../utils.js";
 import * as bootstrap from "bootstrap";
 import * as disc from "../fdc.js";
 import { GoogleDriveLoader } from "../google-drive.js";
 import { toast } from "./toast.js";
 import { errorText } from "./reporting.js";
 import { replaceOrAddExtension } from "../archive.js";
+import { noteEvent } from "./analytics.js";
 
 /**
  * The Google Drive picker: signing in, listing the user's discs, loading one
@@ -138,7 +138,7 @@ export class GoogleDrivePicker {
             dbList.appendChild(row);
             row.querySelector(".name").textContent = item.name;
             row.addEventListener("click", async () => {
-                utils.noteEvent("google-drive", "click", item.name);
+                noteEvent("google-drive", "click", item.name);
                 this.media.setDisc1Image(`gd:${item.id}/${item.name}`);
                 this.modal.hide();
                 try {
