+++ tests/unit/test-utils.js
@@ -1,11 +1,12 @@
 import { describe, it, beforeAll, expect, vi, afterEach } from "vitest";
 
-import { BBC, debounce, getKeyMap, keyCodes, userKeymap } from "../../src/utils.js";
+import { BBC, getKeyMap, keyCodes, userKeymap } from "../../src/utils.js";
 import { ATOM, getKeyMapAtom } from "../../src/utils_atom.js";
 import { processInputParams } from "../../src/url-params.js";
 import { crc32, createZipBlob, replaceOrAddExtension, unzip } from "../../src/archive.js";
 import { readInt16, readInt32, signExtend, stringToUint8Array, uint8ArrayToString } from "../../src/binary.js";
 import { hexbyte, hexword, parseAddr } from "../../src/hex.js";
+import { debounce } from "../../src/debounce.js";
 
 describe("Utils tests", function () {
     describe("parseAddr", function () {
